[Frontend] Share nav item definitions between sidebars and MobileDrawer
